% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readXRF_excel.R
\name{readXRF_excel}
\alias{readXRF_excel}
\title{Function to get XRF analysis data from Excel.}
\usage{
readXRF_excel(file, XRF, ...)
}
\arguments{
\item{file}{The path and file name of the Excel workbook containing the data.}

\item{XRF}{The name of the XRF instrument (e.g. Froya, Thor).}

\item{...}{Any other parameters which will passed directly to read_excel().}
}
\description{
This function will return a tidy data frame from Panalytical E5 XRF
instrument results which are stored in an Excel worksheet.
}
\examples{
readXRF_excel()
}
