% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqrcDBconnect.R
\name{aqrcDBconnect}
\alias{aqrcDBconnect}
\title{Function to connect to the AQRC databases}
\usage{
aqrcDBconnect(network = c("IMPROVE", "CSN"), server)
}
\arguments{
\item{network}{Character string of the network to connect to (IMPROVE or
CSN).}

\item{server}{Character string of the server to connect to. Typically
something like "aqrc-sql", "aqrc-sql-test", and "aqrc-sql-appdev"
(default is the production database server "aqrc-sql").}
}
\value{
A pool connection to the database defined by the parameters.
}
\description{
This function will connect to the AQRC databases (IMPROVE or CSN) and pull
data from multiple tables that are useful for XRF data analysis. The database
connection is done with the help of a config file which is stored in the
../config/config.yml file. See the documentation for the \code{config}
package for details.
}
\details{
This function will open a database connection via the \code{pool}
  package. Because this is opening pool connections to the database, it is
  best practice to ensure these connections are closed after database
  operations are finished. For details of the \code{pool} package see its
  website at
  \href{https://github.com/rstudio/pool}{https://github.com/rstudio/pool}.
}
\examples{
\dontrun{
poolIMP <- aqrcDBconnect("IMPROVE", "aqrc-sql-test")
poolCSN <- aqrcDBconnect("CSN", "aqrc-test")
# Perform some database operations with the connection.
# Once complete, close the pool connection.
pool::poolClose(poolIMP); rm(poolIMP)
pool::poolClose(poolCSN); rm(poolCSN)
}

}
