% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqrcDBconnect.R
\name{aqrcDBdisconnect}
\alias{aqrcDBdisconnect}
\title{Function to disconnect from the database and cleanup the connection}
\usage{
aqrcDBdisconnect(poolConn, pos = NULL)
}
\arguments{
\item{poolConn}{pool object which represents an active connection.}
}
\value{
Nothing.
}
\description{
This function will disconnect from any pool connection and remove the 
connection object from the R environment.
}
\details{
This function will disconnect a database connection via the
  \code{pool} package. After disconnecting it will remove the connection
  object from the R environment.
}
\examples{
# poolIMP is a database connection
aqrcDBdisconnect(poolIMP)
# the connection is closed and the object, poolIMP, is removed.

}
